# packages/backend/src/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cloudflare/check-dns-availability": {
      "get": {
        "operationId": "checkDnsAvailability",
        "parameters": [
          {
            "name": "subdomain",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "domain",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appUrn",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Cloudflare"
        ]
      }
    },
...
    "/api/cloudflare/check-url-availability": {
      "get": {
        "operationId": "checkUrlAvailability",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Cloudflare"
        ]
      }
    },
...
    "/api/cloudflare/remove-catch-all-routes": {
      "post": {
        "operationId": "removeCatchAllRoutes",
        "parameters": [
          {
            "name": "tunnelId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Cloudflare"
        ]
      }
    },
...
  },
  "info": {
    "title": "CI Hub API",
    "description": "API specs for CI Hub",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}